refactor: 명령어로 입력된 position 검증 로직을 커맨드 내부로 이동
